requirements.txt
argon2-cffi==20.1.0
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
astroid==2.4.2
bcrypt==3.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
colorama==0.4.4
dj-database-url==0.5.0
Django @ file:///tmp/build/80754af9/django_1605066529526/work
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-braces==1.14.0
django-multiselectfield==0.1.12
Faker==4.15.0
gunicorn==20.0.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
misaka==2.1.1
Pillow==8.0.1
psycopg2 @ file:///C:/ci/psycopg2_1594305172475/work
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
soupsieve==2.1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
text-unidecode==1.3
toml==0.10.2
whitenoise==5.2.0
wincertstore==0.2
wrapt==1.12.1
